FORTIFY_SOURCE intra-object overflow checking
This adds supporting for detecting buffer overflows from inner objects
for the fortified string family functions. It's comparable to the
_FORTIFY_SOURCE=2 feature in glibc with the additional coverage of
intra-object read overflows for supported functions.

The mem* family functions are left with only the inter-object overflow
checks as is the case with glibc _FORTIFY_SOURCE=2.

This feature is currently hidden behind CONFIG_EXPERT because it's a lot
more likely to uncover benign / intended issues and will need a lot of
runtime testing. It's already useful for finding bugs but it may not yet
be a good idea to use it for hardening unless panics for benign issues
are seen as a lesser evil than the vulnerabilities it can catch.

Signed-off-by: Daniel Micay <danielmicay@gmail.com>
Signed-off-by: Albert I <krascgq@outlook.co.id>