block, bfq-mq, bfq-sq: make log functions print names of calling functions
Add the macro __func__ as a parameter to the invocations of the functions
pr_crit, blk_add_trace_msg and blk_add_cgroup_trace_msg in bfq_log*
functions, in order to include automatically in the log messages
the names of the functions that call the log functions.
The programmer can then avoid doing it.

Signed-off-by: Davide Paganelli <paga.david@gmail.com>
Signed-off-by: Paolo Valente <paolo.valente@linaro.org>
Signed-off-by: Albert I <krascgq@outlook.co.id>