fs/bcachefs: Improve enum-conversion warning fix
On trans_get_update(), the first condition on each update loop before
proceeding to next one is to have iter->btree_id (saved btree_id value) same as
btree_id as it's passed into as an argument.

We can just safely replace the need for explicit cast of btree_id argument by
checking whether the current iter is extents or not, without affecting the
logic.

Signed-off-by: Albert I <kras@raphielgang.org>